+++ fabric21/src/main/java/com/vecoo/extraquests/reward/KeyValueReward.java
@@ -6,6 +6,7 @@ import dev.ftb.mods.ftbquests.quest.Quest;
 import dev.ftb.mods.ftbquests.quest.ServerQuestFile;
 import dev.ftb.mods.ftbquests.quest.reward.Reward;
 import dev.ftb.mods.ftbquests.quest.reward.RewardType;
+import lombok.Getter;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.core.HolderLookup;
@@ -15,6 +16,7 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.server.level.ServerPlayer;
 
+@Getter
 public class KeyValueReward extends Reward {
     public static RewardType TYPE;
 
@@ -64,18 +66,6 @@ public class KeyValueReward extends Reward {
         this.key = buffer.readUtf(Short.MAX_VALUE);
         this.value = buffer.readVarLong();
         this.ignore = buffer.readBoolean();
-    }
-
-    public String getKey() {
-        return this.key;
-    }
-
-    public long getValue() {
-        return this.value;
-    }
-
-    public boolean isIgnore() {
-        return this.ignore;
     }
 
     @Override