+++ forge20/src/main/java/com/vecoo/extraquests/service/QuestTimerService.java
@@ -1,24 +1,24 @@
-package com.vecoo.extraquests.storage;
+package com.vecoo.extraquests.service;
 
 import com.vecoo.extralib.gson.UtilGson;
 import com.vecoo.extralib.task.TaskTimer;
 import com.vecoo.extralib.world.UtilWorld;
 import com.vecoo.extraquests.ExtraQuests;
-import com.vecoo.extraquests.util.Utils;
+import com.vecoo.extraquests.api.factory.ExtraQuestsService;
+import lombok.val;
 import net.minecraft.server.MinecraftServer;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.HashSet;
 import java.util.Set;
 
-public class QuestTimerProvider {
+public class QuestTimerService {
     private transient final String filePath;
     private final Set<QuestTimer> questTimers;
 
-    private transient boolean intervalStarted = false;
     private transient volatile boolean dirty = false;
 
-    public QuestTimerProvider(@NotNull String filePath, @NotNull MinecraftServer server) {
+    public QuestTimerService(@NotNull String filePath, @NotNull MinecraftServer server) {
         this.filePath = UtilWorld.resolveWorldDirectory(filePath, server);
 
         this.questTimers = new HashSet<>();
@@ -54,36 +54,32 @@ public class QuestTimerProvider {
     }
 
     private void saveInternal() {
-        if (!this.intervalStarted) {
-            TaskTimer.builder()
-                    .withoutDelay()
-                    .interval(75 * 20L)
-                    .infinite()
-                    .consume(task -> {
-                        if (ExtraQuests.getInstance().getServer().isRunning() && this.dirty) {
-                            UtilGson.writeFileAsync(this.filePath, "quest_timers.json",
-                                    UtilGson.getGson().toJson(this)).thenRun(() -> this.dirty = false);
-                        }
-                    })
-                    .build();
-
-            this.intervalStarted = true;
-        }
+        TaskTimer.builder()
+                .withoutDelay()
+                .interval(150 * 20L)
+                .infinite()
+                .consume(task -> {
+                    if (ExtraQuests.getInstance().getServer().isRunning() && this.dirty) {
+                        UtilGson.writeFileAsync(this.filePath, "quest_timers.json",
+                                UtilGson.getGson().toJson(this)).thenRun(() -> this.dirty = false);
+                    }
+                })
+                .build();
     }
 
     public void init() {
         this.questTimers.clear();
 
         UtilGson.readFileAsync(this.filePath, "quest_timers.json", el -> {
-            QuestTimerProvider questTimerProvider = UtilGson.getGson().fromJson(el, QuestTimerProvider.class);
-            long time = System.currentTimeMillis();
+            val questTimerProvider = UtilGson.getGson().fromJson(el, QuestTimerService.class);
+            val time = System.currentTimeMillis();
 
             for (QuestTimer questTimer : questTimerProvider.getQuestTimers()) {
                 if (questTimer.endTime() > time) {
                     this.questTimers.add(questTimer);
-                    Utils.startQuestTimer(questTimer);
+                    ExtraQuestsService.startQuestTimer(questTimer);
                 } else {
-                    Utils.questReset(questTimer);
+                    ExtraQuestsService.questReset(questTimer);
                 }
             }
         }).join();