+++ changelog.md
@@ -1,1 +1,3 @@
-* Fix storage (Forge 1.20.1).
+* Change type config json to yml.
+* Checking reload config and remove reload storage.
+* Improve code.